Perms update inline with 4.4.4 webclient see #9663
Uses request.session['user_id'] and ['group_id'] to keep track
of current group and user.

Add decorators that can look-up the wrapper objects and add them
to the page context for use in the template as ome.active_user
and ome.active_group.
Update pages to use this @render_response decorator.